# swagger.json
{
    "swagger": "2.0",
    "info": {
            "version": "1.0.0",
            "title": "FamilySize Servicio Web",
            "description": "Servicio Web para la obtención del menú de FamilySize Restaurant",
            "license": {
            "name": "ISC",
            "url": "https://opensource.org/licenses/ISC"
        }
    },
    "host": "familysize-servicio-web.herokuapp.com",
    "basePath": "/",
    "tags": [
        {
            "name": "Menú",
            "description": "Obtención de los items del menú"
        }
    ],
    "schemes": ["https"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
        "/menu": {
            "get": {
                "tags": ["Menú"],
                "summary": "Get para obtener todos los items del menú",
                "responses": {
                    "200": {
                        "description": "Arreglo de items del menu",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "ok": {
                                    "type": "boolean"
                                },
                                "menu": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Item"
                                    }
                                }
                            }
                        }
                    },
                    "500" : {
                        "description": "Ocurrió un error en el servidor",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/menu/{id}": {
            "get": {
                "tags": ["Menú"],
                "summary": "Get para obtener un item particular del menú",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        },
                        "description": "Identificador unívoco del item"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item identificado por id",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "ok": {
                                    "type": "boolean"
                                },
                                "item": {
                                    "$ref": "#/definitions/Item"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se logró encontrar un item con el id suministrado",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500" : {
                        "description": "Ocurrió un error en el servidor",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/menu/{id}/picture": {
            "get": {
                "tags": ["Menú"],
                "summary": "Get para obtener la imagen de un item particular del menú",
                "produces": "image/jpeg",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        },
                        "description": "Identificador unívoco del item"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Imagen del menú seleccionado",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description":"No se logró encontrar un item con el id suministrado",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ocurrió un error en el servidor",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Item": {
            "properties": {
                "id": {
                    "type": "integer",
                    "uniqueItems": true
                },
                "name": {
                    "type": "string"
                },
                "cuisine": {
                    "type": "string"
                },
                "is_vege": {
                    "type": "boolean"
                },
                "is_vegan": {
                    "type": "boolean"
                },
                "is_coeliac": {
                    "type": "boolean"
                },
                "has_alcohol": {
                    "type": "boolean"
                },
                "cost": {
                    "type": "number"
                }
            }
        },
        "ErrorResponse": {
            "properties": {
                "ok": {
                    "type": "boolean"
                },
                "error": {
                    "type": "string"
                }
            }
        }
    }
}